% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Function to create a map for a given state
and given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer represeting the state's number.}

\item{year}{Integer for the year.}
}
\value{
a map plot.
}
\description{
This functin takes a state number and a year and
creates a map for the subset of data corresponding
to those values.
}
