% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads the csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Character with the path to the file.}
}
\value{
A tibble dataframe.
}
\description{
Reads the csv file with the data from
the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System.
}
\details{
If filename is not found then stops.
}
