% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads files for particular years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Numeric (integer) vector.}
}
\value{
List of tibbles, each for one element in \code{years}.
}
\description{
\code{fars_read_years} takes a numeric vector with the
years to read.
}
\details{
For any element in \code{years} if this element is an
invalid year, the function returns \code{NULL}.
}
\seealso{
\link[=fars_read]{make_filename}.
}
