% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a summary for a set of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Numeric (integer) vector with the years
you want to summarize.}
}
\value{
tibble with the summarized data
}
\description{
This functions takes a numeric vector
that contains the years that you want
to summarize
}
\seealso{
\link{fars_read_years}
}
