% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numeric vector}
}
\value{
A character vector following the pattern
"accident_\%d.csv.bz2"
}
\description{
\code{make_filename} creates a character vector using
a numeric vector with the years.
}
\section{Warning}{
 Non-integer numbers are converted to integers.
}

\examples{
make_filename(2010)
make_filename(c(2010,2020))
}
