% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads the csv file}
\usage{
fars_read(filename)
}
\description{
\code{fars_read} reads the csv file with the data from
the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System.

@param  filename Character with the path to the file.

@details If \code{filename} is not found then stops.

@return A tibble dataframe

@examples
\dontrun{
 fars_read("path/to/file")
 }

@import readr
@import dplyr
}
